django_api/requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
